+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# v 3.14.0 (?)
+# v 3.14.0 (2025-08-26)
 Changes in this release:
 - Allow to configure `PIP_CONSTRAINT` environment variable on the remote host during project install, using the `--pip-constraint` pytest option.
 - Fix format of inmanta config saved in the project synced to remote orchestrator.